Add benchmarks
The LTW benchmarks don't mean a lot yet (it's hard to change the
number of aspects at runtime with AspectJ). The Spring AOP numbers
are startling - the startup time is driven by the number of
pointcuts having @aspect expressions (if you remove the @matcher
the startup time is the same as with 0 aspects), not particularly
by the number of beans.